Favors Plug instead of Pathogen
Maybe this is a bad idea. Keeping all these plugins on this repository
will be history, so I am not afraid of not finding this code in the
future, as well as messing with Git modules and managing them.

The problem is that an internet connection and Git is now a requirement
for this. Let's see if this will bother me in the future.